+++ config/config.go
@@ -637,6 +637,16 @@ func (root *Root) IsTerragruntChangeDetectionEnabled() bool {
 	}
 }
 
+// IsTargetsEnabled returns the configured `terramate.config.targets.enabled` option.
+func (root *Root) IsTargetsEnabled() bool {
+	if root.tree.Node.Terramate != nil &&
+		root.tree.Node.Terramate.Config != nil &&
+		root.tree.Node.Terramate.Config.Targets != nil {
+		return root.tree.Node.Terramate.Config.Targets.Enabled
+	}
+	return false
+}
+
 // Skip returns true if the given file/dir name should be ignored by Terramate.
 func Skip(name string) bool {
 	// assumes filename length > 0